Muraraba iyo LoginActivity.java n akajagarindacariko ndayikora fresh gusa uwushobora akosore

+++ app/src/main/java/com/skilledhacker/developer/musiqx/Utilities/Utilitties.java
@@ -61,7 +61,7 @@ public class Utilitties {
      * */
     public int progressToTimer(int progress, int totalDuration) {
         int currentDuration = 0;
-        totalDuration = (int) (totalDuration / 1000);
+        totalDuration = totalDuration / 1000;
         currentDuration = (int) ((((double)progress) / 100) * totalDuration);
 
         // return current duration in milliseconds
